Add `logging#Config`
